Update Keyboard.lua
AreKeysDown() and AreEitherKeysDown would be more convenient if it supported arbitrary amount of keycodes, instead of it being constrained to 2. Thus, this change makes these two methods expect a table of keycodes as an argument instead of 2 keycodes separately. 

ALTERNATIVELY, I could handle the keycodes as variadic arguments so there is no need to explicitly feed in the keycodes as a table, which would be compatible with current design of directly feeding in the keycodes as multiple arguments.